requirements.txt
# Python 3.11
# Core data pipeline
faker>=24.0.0
pandas>=2.2.0
pyarrow>=15.0.0
psycopg2-binary>=2.9.9
boto3>=1.34.0          # MinIO / S3 upload (s1_generate_raw)

# Airflow — TaskFlow API (only needed if running the DAG, not required for running scripts directly)
apache-airflow>=2.9.0

# Config / environment
python-dotenv>=1.0.0

# Utilities (stdlib supplements)
numpy>=1.26.0          # feature normalization in Gold layer
scikit-learn>=1.4.0    # k-means fallback when MADlib not accessible
